requirements.txt
flask>=3.0
flask-login>=0.6
flask-wtf>=1.2
bcrypt>=4.0
pyyaml>=6.0
python-dotenv>=1.0
gunicorn>=22.0
gevent>=24.0
paramiko>=3.0
